backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==3.2.25
django-cors-headers==4.4.0
djangorestframework==3.15.1
filelock==3.15.4
fsspec==2024.6.1
gunicorn==22.0.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sqlparse==0.5.0
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.12.2
urllib3==2.2.2
matplotlib
wordcloud